Add cast to int to silence format mismatch warning
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10639 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_vmd/mod_vmd.c
@@ -282,7 +282,7 @@ static void find_beep(vmd_session_info_t *vmd_info, switch_frame_t *frame)
 			}
 
 			switch_event_add_header_string(event, SWITCH_STACK_BOTTOM, "Beep-Status", "stop");
-			switch_event_add_header(event, SWITCH_STACK_BOTTOM, "Beep-Time", "%d", vmd_info->timestamp / POINTS);
+			switch_event_add_header(event, SWITCH_STACK_BOTTOM, "Beep-Time", "%d", (int)vmd_info->timestamp / POINTS);
 			switch_event_add_header(event, SWITCH_STACK_BOTTOM, "Unique-ID",
 				"%s", switch_core_session_get_uuid(vmd_info->session));
 			switch_event_add_header(event, SWITCH_STACK_BOTTOM, "Frequency", "%6.4lf", vmd_info->beep_freq);
